+++ client/cypress/component/groundwater/Groundwater.cy.js
@@ -1,16 +1,16 @@
-import GroundWater from "@/components/groundwater/GroundWater.vue";
+import GroundWaterQuality from "@/components/groundwater/GroundWaterQuality.vue";
 import groundWaterStations from '../../fixtures/groundWaterStations.json';
 
 const pointCount = groundWaterStations.features.length;
 
-describe('<GroundWater />', () => {
+describe('<GroundWaterQuality />', () => {
     beforeEach(() => {
         cy.intercept('**/stations', { fixture: 'groundWaterStations.json' });
         cy.intercept('**/report', { fixture: 'groundWaterChemistry.json' });
     });
 
     it('mounts and loads main page contents', () => {
-        cy.mount(GroundWater);
+        cy.mount(GroundWaterQuality);
         cy.get('.mapboxgl-canvas').should('exist').and('be.visible')
         // zoom out of the map, showing all points
         cy.wait(1000);
@@ -19,7 +19,7 @@ describe('<GroundWater />', () => {
         cy.get('.map-point-count > div > i').should('contain', pointCount);
     });
     it('mounts and loads report contents', () => {
-        cy.mount(GroundWater);
+        cy.mount(GroundWaterQuality);
         cy.get('.q-virtual-scroll__content > .q-item:first').click();
         // details are displayed
         cy.get('.selected-point').should('not.be.empty');
